+++ aiohttp_auth/auth/ticket_auth.py
@@ -44,8 +44,8 @@ class TktAuthentication(AbstractAuthentication):
         self._ticket = TicketFactory(secret)
         self._max_age = max_age
         if (self._max_age is not None and
-            reissue_time is not None and
-            reissue_time < self._max_age):
+                reissue_time is not None and
+                reissue_time < self._max_age):
             self._reissue_time = max_age - reissue_time
         else:
             self._reissue_time = None
@@ -106,10 +106,10 @@ class TktAuthentication(AbstractAuthentication):
 
             # Check if we need to reissue a ticket
             if (self._reissue_time is not None and
-                now >= (fields.valid_until - self._reissue_time)):
-
+                    now >= (fields.valid_until - self._reissue_time)):
                 # Reissue our ticket, and save it in our request.
-                request[_REISSUE_KEY] = self._new_ticket(request, fields.user_id)
+                request[_REISSUE_KEY] = self._new_ticket(request,
+                                                         fields.user_id)
 
             return fields.user_id
 
@@ -117,13 +117,13 @@ class TktAuthentication(AbstractAuthentication):
             return None
 
     async def process_response(self, request, response):
-        """If a reissue was requested, only reiisue if the response was a
+        """If a reissue was requested, only reissue if the response was a
         valid 2xx response
         """
         if _REISSUE_KEY in request:
-            if (response.started or
-                not isinstance(response, web.Response) or
-                response.status < 200 or response.status > 299):
+            if (response.prepared or
+                    not isinstance(response, web.Response) or
+                    response.status < 200 or response.status > 299):
                 return
 
             await self.remember_ticket(request, request[_REISSUE_KEY])